feat(unity_builder.yaml): Auto-release & Versioning
First now this workflow is only executed when we push on main from a branch that have a name as 'release-1.0.0' for example.
Some debug comes like display all files contains in build directory of artifacts.
When the workflow success, you can go in the Releases part of the repository and see you new release with a 'Release.zip' files that contains all build of our projects in all version like 'Windows, Linux, Xbox...).

Notion #23