fix(ShopData): Testing FillShopCommand
During the test of the command 'FillShop', I see that the undo command of the fill shop is causing some trouble.
Why ? Because, the reference already store the same value as current shop and old shop.
For this problem, I add a 'DeepCopy' function in 'ShopData' and 'ShopSection' that is call in the 'FillShop' command for create a copy of the current shop and store it as en old.