fix(CommandSystem): Rework command system
Rework the command pattern in the OMGG.DesignPattern, for having a state when we execute a command, for knowing the successfull state of the last command.
Rework the BuyCommand for network and non network compatibilty.

Now the 'PlayerController', can buy a card and send in network the the command only if he can bought it.